// src/components/modules/NFTDetails/types.ts
export interface NFTDetailsTypes {
  amount?: string;
  image?: string;
  contract_type?: string;
  metadata: string;
  name?: string;
  information?: object;
  owner_of?: string;
  symbol?: string;
  token_address?: string;
  token_hash?: string;
  token_id?: string | number;
  token_uri?: string;
  description?: string;
}